backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
sentence-transformers==3.1.1
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.9.2
python-dotenv==1.0.1
